requirements.txt
Flask==1.1.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-cloud-pubsub==2.4.0
google-cloud-storage